projects/AccountTracker/requirements.txt
django
psycopg2
django-localflavor
social-auth-app-django
